Rebase code + migration to TypeScript (initial commit)
