+++ const/const.ts
@@ -1,0 +1,12 @@
+
+// @ts-ignore
+const LOG_INFO = 1,
+      LOG_ERROR = 2,
+      PREFIX = '!instant'
+
+
+module.exports = {
+    LOG_INFO,
+    LOG_ERROR,
+    PREFIX
+}
